requirements.txt
aioquic==1.2.0
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
appdirs==1.4.4
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
azure-communication-callautomation==1.3.0b1
azure-core==1.32.0
azure-eventgrid==4.11.0
beautifulsoup4==4.9.3
blinker==1.8.2
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
cloudscraper==1.2.71
conda @ file:///croot/conda_1729193882500/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
constantly==23.10.4
cryptography @ file:///croot/cryptography_1724940562255/work
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
distro @ file:///croot/distro_1714488253808/work
elastic-transport==8.15.1
elasticsearch==8.16.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
frozendict @ file:///croot/frozendict_1713194832637/work
future==1.0.0
geocoder==1.38.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.5.0
incremental==24.7.2
isodate==0.7.2
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
kaitaistruct==0.10
ldap3==2.9.1
libmambapy @ file:///croot/mamba-split_1725563367192/work/libmambapy
lxml==5.3.0
MarkupSafe==3.0.2
menuinst @ file:///croot/menuinst_1723567589013/work
mitmproxy==11.0.0
mitmproxy_rs==0.9.3
msgpack==1.1.0
msrest==0.7.1
numpy==2.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.2.3
parsel==1.9.1
passlib==1.7.4
pillow==11.0.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
Protego==0.3.1
protobuf==5.28.2
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyDispatcher==2.0.7
pyee==11.1.1
pylsqpack==0.3.18
pyOpenSSL==20.0.1
pyparsing==3.1.4
pyperclip==1.9.0
pyppeteer==2.0.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
pytz==2024.2
queuelib==1.7.0
ratelim==0.1.6
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
Scrapy==2.11.2
selenium==4.26.1
service-identity==24.2.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tldextract==5.1.3
tornado==6.4.1
tqdm @ file:///croot/tqdm_1724853939799/work
trio==0.27.0
trio-websocket==0.11.1
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
Twisted==24.10.0
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.6
urllib3==1.26.20
urwid==2.6.15
w3lib==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.2
wheel==0.44.0
wsproto==1.2.0
zipp==3.21.0
zope.interface==7.1.1
zstandard @ file:///croot/zstandard_1728569189425/work
gunicorn